Disable management of all interface to avoid collision with cloud-init
cloud-init takes currently over control of all ethernet interfaces on a
device so that it will conflict with network-manager who wants to
control them too. As long as we're missing detection of files in
/etc/network/interfaces.d (that is where cloud-init write the configuration
files to) and we have way to disabled clout-init's managed of ethernet
devices we disabled support for those.