refactor run-tests.sh to match review comments
The run-tests.sh script does not remove the tests-extras repository
when the tests are finished. Instead it reuse it but first make sure
that it is in a know state by hard reset and pulling.

Moreover the tests-extras is now a hidden directory and the git is
told not to track the .tests-extras. The .gitignore is commited but
only when it is not locally modfied prior to running this script.