Remove assertion for empty DHCP6 options
It turns out that some routers return responses to Information-request
messages that do not contain any of the options that we insert in the
"options" table. When that happened and the info-only flag for DHCP6
was set, the assertion was triggered and NetworkManager crashed. We
remove the assertion as having empty options is a possibility and is
harmless anyway.