Change debug activation to avoid having to re-start NM
To be able to re-start NM the core-support plug was being used, which
would have never been accepted in the Ubuntu store, as that interface
is meant to be used by the core snap only. Instead, avoid having to
restart by using nmcli to dynamically activate/deactivate debug option.
For this we need to use the nmcli plug instead for the configure hook.