Remove calls to set log level in wrapper
The calls were actually failing as NM was not started yet. Furthermore,
the log level is set via a call option in NM. Also remove nmcli-internal
as after this it is not used anymore and probably there will be no need
for it in the future (we restart always NM when the settings change
instead of dynamically changing it).